Reorganize frontend tests.
