+++ frontend/src/types/index.ts
@@ -1,7 +1,6 @@
 export type * from "./api/api";
 export type * from "./api/errors";
 export type * from "./api/events";
-export type * from "./api/fetch";
 export type * from "./props";
 
 export type Maybe<T> = T | null | undefined;
