+++ frontend/test/mocks/backend.ts
@@ -1,4 +1,4 @@
 import { setupServer } from "msw/node";
-import { handlers } from "./handlers";
+import { handlers } from "./handlers.ts";
 
 export const backend = setupServer(...handlers);
