+++ frontend/test/store/slices/channels.spec.ts
@@ -1,9 +1,9 @@
 import { createEntityAdapter } from "@reduxjs/toolkit";
 import { waitFor } from "@testing-library/react";
 import { http, HttpResponse } from "msw";
-import { backend } from "#/mocks/backend";
-import { renderHookWithProviders } from "#/utils/test-utils";
-import { appendChannels, useListChannelsQuery } from "./channels";
+import { appendChannels, useListChannelsQuery } from "#/store/slices/channels";
+import { backend } from "../../mocks/backend";
+import { renderHookWithProviders } from "../../utils/test-utils";
 import type { Channel } from "#/api";
 
 describe("channelsApi", () => {
