+++ frontend/vitest.setup.ts
@@ -1,5 +1,5 @@
 import "@testing-library/jest-dom";
-import { backend } from "#/mocks/backend";
+import { backend } from "./test/mocks/backend";
 
 beforeAll(() => {
   backend.listen({
